+++ crates/cast/tests/e2e/account/mod.rs
@@ -1,3 +1,4 @@
 mod add;
 mod create;
 mod deploy;
+mod delete;